- Modified AttributeLock lock/unlock functions to work with new fieldset implementation - Fixed uniqueness error when unlocking attributes

+++ projects/views.py
@@ -33,7 +33,7 @@ from rest_framework_extensions.mixins import NestedViewSetMixin
 
 from projects.exporting.document import render_template
 from projects.exporting.report import render_report_to_response
-from projects.helpers import DOCUMENT_CONTENT_TYPES, get_file_type, TRUE
+from projects.helpers import DOCUMENT_CONTENT_TYPES, get_file_type, TRUE, get_attribute_lock_data
 from projects.importing import AttributeImporter, AttributeUpdater
 from projects.models import (
     FieldComment,
@@ -846,19 +846,6 @@ class AttributePagination(pagination.PageNumberPagination):
     page_size = 2000
 
 
-def get_attribute_lock_data(attribute_identifier):
-    if "[" in attribute_identifier and "]." in attribute_identifier:
-        fieldset_attribute_identifier = attribute_identifier.split("[")[0]
-        fieldset_attribute_index = attribute_identifier.split("[")[1].split("].")[0]
-        attribute_identifier = attribute_identifier.split("].")[1]
-        return {
-            "attribute_identifier": attribute_identifier,
-            "fieldset_attribute_identifier": fieldset_attribute_identifier,
-            "fieldset_attribute_index": fieldset_attribute_index
-        }
-    return {"attribute_identifier": attribute_identifier}
-
-
 class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Attribute.objects.all()
     serializer_class = SimpleAttributeSerializer
@@ -902,19 +889,24 @@ class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
             fieldset_attribute_index=attribute_lock_data.get("fieldset_attribute_index")
         ).delete()
 
-        attribute_lock = AttributeLock.objects.filter(
-            project=project,
-            attribute=attribute,
-            fieldset_attribute=fieldset_attribute,
-            fieldset_attribute_index=attribute_lock_data.get("fieldset_attribute_index")
-        ).first()
+        if not fieldset_attribute:
+            attribute_lock = AttributeLock.objects.filter(
+                project=project,
+                attribute=attribute
+            ).first()
+        else:
+            attribute_lock = AttributeLock.objects.filter(
+                project=project,
+                fieldset_attribute=fieldset_attribute,
+                fieldset_attribute_index=attribute_lock_data.get("fieldset_attribute_index")
+            ).first()
 
         if attribute_lock and (datetime.now(timezone.utc) - attribute_lock.timestamp).total_seconds() >= 900:  # 15 min
             attribute_lock.delete()
             attribute_lock = None
 
         if not attribute_lock:
-            attribute_lock = AttributeLock.objects.create(
+            attribute_lock, created = AttributeLock.objects.get_or_create(
                 project=project,
                 attribute=attribute,
                 fieldset_attribute=fieldset_attribute,
@@ -955,16 +947,21 @@ class AttributeViewSet(viewsets.ReadOnlyModelViewSet):
 
             attribute_lock_data = get_attribute_lock_data(request.data["attribute_identifier"])
 
-            attribute_lock = AttributeLock.objects.get(
-                project__name=project_name,
-                attribute__identifier=attribute_lock_data.get("attribute_identifier"),
-                fieldset_attribute__identifier=attribute_lock_data.get("fieldset_attribute_identifier"),
-                fieldset_attribute_index=attribute_lock_data.get("fieldset_attribute_index"),
-                user=request.user,
-            )
-            attribute_lock.delete()
-        except AttributeLock.DoesNotExist:
-            pass  # No attribute to unlock, request was still successful
+            if not attribute_lock_data.get("fieldset_attribute"):
+                attribute_lock = AttributeLock.objects.filter(
+                    project__name=project_name,
+                    attribute__identifier=attribute_lock_data.get("attribute_identifier"),
+                    user=request.user,
+                )
+            else:
+                attribute_lock = AttributeLock.objects.filter(
+                    project__name=project_name,
+                    fieldset_attribute__identifier=attribute_lock_data.get("fieldset_attribute_identifier"),
+                    fieldset_attribute_index=attribute_lock_data.get("fieldset_attribute_index"),
+                    user=request.user,
+                )
+            if attribute_lock:
+                attribute_lock.delete()
         except Exception as exc:
             log.error(exc)
             return HttpResponse(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
